# Add option to use Wasmtime caching