--- a/Cargo.toml
+++ b/Cargo.toml
@@ -4,8 +4,6 @@ version = "0.1.0"
 authors = ["Matt Butcher <matt.butcher@microsoft.com>"]
 edition = "2018"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
 hyper = "0.13"
 tokio = { version = "0.2", features = ["full"] }
@@ -13,7 +11,8 @@ futures = "0.3"
 anyhow = "1.0"
 toml = "0.5"
 serde = { version = "1.0", features = ["derive"] }
-wasmtime = "0.20"
-wasmtime-wasi = "0.20"
-wasi-common = "0.20"
+wasmtime = "0.22"
+wasmtime-wasi = "0.22"
+wasmtime-cache = "0.22"
+wasi-common = "0.22"
 clap = "2.33.3"