ref #3 

This commit updates WAGI to allow using the Wasmtime cache in order to speed up the execution of modules.
(more info about Wasmtime caching - https://docs.wasmtime.dev/cli-cache.html)

Specifically, for large modules, this has a noticeable impact on instantiation time.
While I don't have extensive benchmarks yet, for a 2.1M module (built from https://github.com/deislabs/env_wagi):

- without caching:

```
cargo run --release -- --config examples/modules.toml                            
    Finished release [optimized] target(s) in 0.14s
     Running `target/release/wagi --config examples/modules.toml`
=> Starting server
instantiation time for module examples/env_wagi.wasm: 36.88375ms
Setting status to 200 OK
instantiation time for module examples/env_wagi.wasm: 36.364857ms
Setting status to 200 OK
instantiation time for module examples/env_wagi.wasm: 34.18837ms
Setting status to 200 OK
instantiation time for module examples/env_wagi.wasm: 33.508107ms
```

- with caching:

```
cargo run --release -- --config examples/modules.toml --cache examples/cache.toml
    Finished release [optimized] target(s) in 0.14s
     Running `target/release/wagi --config examples/modules.toml --cache examples/cache.toml`
=> Starting server
instantiation time for module examples/env_wagi.wasm: 22.336185ms
Setting status to 200 OK
instantiation time for module examples/env_wagi.wasm: 21.358911ms
Setting status to 200 OK
instantiation time for module examples/env_wagi.wasm: 16.510206ms
Setting status to 200 OK
instantiation time for module examples/env_wagi.wasm: 17.579882ms
Setting status to 200 OK
instantiation time for module examples/env_wagi.wasm: 15.873305ms
```

I suspect the improvements in performance are better for larger modules, but this is something we should test further.